+++ code/packages/typescript/sir-runtime-core/src/index.ts
@@ -31,6 +31,7 @@ export {
   globalGet,
   globalGetStatic,
   print,
+  puts,
   callBuiltin,
   builtinClosure,
   doubleSplatMerge,